requirements.txt
numpy==1.16.2
opencv-python==4.1.0.25
python-constraint==1.4.0
